libnbcompat: force lchmod replacement on Linux
Linux doesn't support changing the mode of symlinks, so lchmod
always returns EOPNOTSUPP. AC_REPLACE_FUNCS is able to detect glibc
stubs for this function (among others), but not on musl. This causes
failures when tools, such as pax(1), attempt to preserve the
modification times of symlinks.

To resolve this, set the configure cache variable to always replace
lchmod on Linux.