firefox: fix build with rust 1.45
rust 1.45 added an embed-bitcode flag which has some compatibility
issues with LTO[0]. This causes build failures in firefox like the
following:

  6:43.34 error: options `-C embed-bitcode=no` and `-C lto` are incompatible
  6:43.49 error: could not compile `gkrust`.

This is already fixed in upstream firefox in commits [1] and [2].
The second reverts the first in favor of a slightly different fix.
The -Cembed-bitcode=yes hunk is the only one actually necessary to
fix the build failure, so only that one file is patched.

[0] https://github.com/rust-lang/rust/blob/f173a4b/RELEASES.md#compatibility-notes
[1] https://hg.mozilla.org/mozilla-central/rev/bb79da633a2efe12f7a0a6e89d925961b8b444a7
[2] https://hg.mozilla.org/mozilla-central/rev/e5d2a6d5187b8f1420b65089c4d841ec29a42897